refactor(settings): split settings owner modules
